initial creation of RandBLAS/testing
