+++ test/comparison.hh
@@ -29,9 +29,9 @@
 
 #pragma once
 
+#include "RandBLAS/testing/comparison.hh"
 #include "RandBLAS.hh"
 #include <gtest/gtest.h>
-#include <cmath>
 #include <cmath>
 #include <numeric>
 #include <iostream>
@@ -42,56 +42,19 @@ namespace test::comparison {
 using blas::Layout;
 using blas::Op;
 
-/** Tests two floating point numbers for approximate equality.
- * See https://randomascii.wordpress.com/2012/02/25/comparing-floating-point-numbers-2012-edition/
- *
- * @param[in] A    one number to compare
- * @param[in] B    the second number to compare
- * @param[in] atol is an absolute tolerance that comes into play when
- *                 the values are close to zero
- * @param[in] rtol is a relative tolerance, which should be close to
- *                 epsilon for the given type.
- * @param[inout] str a stream to send a decritpive error message to
- *
- * @returns true if the numbers are atol absolute difference or rtol relative
- *          difference from each other.
- */
-template <typename T>
-bool approx_equal(T A, T B, std::ostream &str,
-    T atol = T(10)*std::numeric_limits<T>::epsilon(),
-    T rtol = std::numeric_limits<T>::epsilon())
-{
-    // Check if the numbers are really close -- needed
-    // when comparing numbers near zero.
-    T diff_ab = abs(A - B);
-    if (diff_ab <= atol)
-        return true;
-
-    T max_ab = std::max(abs(B), abs(A));
-
-    if (diff_ab <= max_ab * rtol)
-        return true;
-
-    str.precision(std::numeric_limits<T>::max_digits10);
-
-    str << A << " != " << B << " with absDiff=" << diff_ab
-        << ", relDiff=" << max_ab*rtol << ", atol=" << atol
-        << ", rtol=" << rtol;
-
-    return false;
-}
+// TODO: Make macros that can automatically inject __PRETTY_FUNCTION__, __FILE__, and __LINE__ into the calls below.
+//       This is slightly complicated by the presence of optional arguments in the function definition.
 
 template <typename T>
 void approx_equal(T A, T B, const char* testName, const char* fileName, int lineNo, T atol, T rtol) {
     std::ostringstream oss;
-    if (!approx_equal(A, B, oss, atol, rtol)) {
+    if (!RandBLAS::testing::approx_equal(A, B, oss, atol, rtol)) {
         FAIL() << std::endl << fileName << ":" << lineNo << std::endl
             << testName << std::endl << "Test failed. " << oss.str() << std::endl;
         oss.str("");
     }
     return;
 }
-
 
 
 /** Test two arrays are approximately equal elementwise.
@@ -102,7 +65,7 @@ void approx_equal(T A, T B, const char* testName, const char* fileName, int line
  * @param[in] testName the name of the test, used in decriptive message
  * @param[in] fileName the name of the file, used in descriptive message
  * @param[in] lineNo the line tested, used in descriptive message
- * 
+ *
  * aborts if any of the elemnts are not approximately equal.
  */
 template <typename T>
@@ -118,7 +81,7 @@ void buffs_approx_equal(
 ) {
     std::ostringstream oss;
     for (int64_t i = 0; i < size; ++i) {
-        if (!approx_equal(actual_ptr[i], expect_ptr[i], oss, atol, rtol)) {
+        if (!RandBLAS::testing::approx_equal(actual_ptr[i], expect_ptr[i], oss, atol, rtol)) {
             FAIL() << std::endl << fileName << ":" << lineNo << std::endl
                 << testName << std::endl << "Test failed at index " << i
                 << " " << oss.str() << std::endl;
@@ -126,9 +89,6 @@ void buffs_approx_equal(
         }
     }
 }
-
-// TODO: Make macros that can automatically inject __PRETTY_FUNCTION__, __FILE__, and __LINE__ into the calls below.
-//       This is slightly complicated by the presence of optional arguments in the function definition.
 
 template <typename T>
 void buffs_approx_equal(
@@ -145,7 +105,7 @@ void buffs_approx_equal(
 ) {
     std::ostringstream oss;
     for (int64_t i = 0; i < size; ++i) {
-        if (!approx_equal(actual_ptr[i*inc_actual], expect_ptr[i*inc_expect], oss, atol, rtol)) {
+        if (!RandBLAS::testing::approx_equal(actual_ptr[i*inc_actual], expect_ptr[i*inc_expect], oss, atol, rtol)) {
             FAIL() << std::endl << fileName << ":" << lineNo << std::endl
                 << testName << std::endl << "Test failed at index " << i
                 << " " << oss.str() << std::endl;
@@ -206,7 +166,7 @@ void matrices_approx_equal(
     if (transB == blas::Op::NoTrans) {
         for (int64_t i = 0; i < m; ++i) {
             for (int64_t j = 0; j < n; ++j) {
-                if (!approx_equal(A[idxa(i, j)], B[idxb(i, j)], oss, atol, rtol)) {
+                if (!RandBLAS::testing::approx_equal(A[idxa(i, j)], B[idxb(i, j)], oss, atol, rtol)) {
                     FAIL() << std::endl << fileName << ":" << lineNo << std::endl
                         << testName << std::endl << "\tTest failed at index ("
                         << i << ", " << j << ")\n\t" << oss.str() << std::endl;
@@ -217,7 +177,7 @@ void matrices_approx_equal(
     } else {
         for (int64_t i = 0; i < m; ++i) {
             for (int64_t j = 0; j < n; ++j) {
-                if (!approx_equal(A[idxa(i, j)], B[idxb(j, i)], oss, atol, rtol)) {
+                if (!RandBLAS::testing::approx_equal(A[idxa(i, j)], B[idxb(j, i)], oss, atol, rtol)) {
                     FAIL() << std::endl << fileName << ":" << lineNo << std::endl
                         << testName << std::endl << "\tTest failed at index ("
                         << j << ", " << i << ")\n\t"  << oss.str() << std::endl;
