+++ test/test_matmul_cores/linop_common.hh
@@ -29,243 +29,28 @@
 
 #pragma once
 
-#include "RandBLAS/config.h"
-#include "RandBLAS/base.hh"
-#include "RandBLAS/dense_skops.hh"
-#include "RandBLAS/sparse_skops.hh"
-#include "RandBLAS/skge.hh"
-#include "RandBLAS/sparse_data/spmm_dispatch.hh"
-#include "RandBLAS/util.hh"
+#include "RandBLAS/testing/linop_common.hh"
 #include "test/comparison.hh"
-#include <functional>
-#include <vector>
-#include <tuple>
-#include <numeric>
-#include <algorithm>
 
 
 namespace test::linop_common {
 
 using blas::Layout;
 using blas::Op;
-using RandBLAS::sparse_data::COOMatrix;
-using RandBLAS::sparse_data::CSRMatrix;
-using RandBLAS::sparse_data::CSCMatrix;
-
-#ifdef __cpp_concepts
-using RandBLAS::SparseMatrix;
-#else
-#define SparseMatrix typename
-#endif
-using RandBLAS::SparseSkOp;
-using RandBLAS::DenseSkOp;
 using RandBLAS::RNGState;
-using RandBLAS::DenseDist;
-using RandBLAS::dims_before_op;
 using RandBLAS::offset_and_ldim;
-using RandBLAS::layout_to_strides;
-using RandBLAS::dims64_t;
 
+using RandBLAS::testing::reference_left_apply;
+using RandBLAS::testing::left_apply;
+using RandBLAS::testing::reference_right_apply;
+using RandBLAS::testing::right_apply;
 
-template <typename T>
-std::vector<T> eye(int64_t n) {
-    std::vector<T> A(n * n, 0.0);
-    for (int i = 0; i < n; ++i)
-        A[i + n*i] = 1.0;
-    return A;
-}
+using RandBLAS::testing::to_explicit_buffer;
+using RandBLAS::testing::dimensions;
 
-template <typename T, typename RNG=r123::Philox4x32>
-auto random_matrix(int64_t m, int64_t n, RNGState<RNG> s) {
-    std::vector<T> A(m * n);
-    DenseDist DA(m, n);
-    auto next_state = RandBLAS::fill_dense(DA, A.data(), s);
-    std::tuple<std::vector<T>, Layout, RNGState<RNG>> t{A, DA.natural_layout, next_state};
-    return t;
-}
+using RandBLAS::testing::random_matrix;
+using RandBLAS::testing::eye;
 
-template <typename LINOP>
-dims64_t dimensions(LINOP &S) {return {S.n_rows, S.n_cols};}
-
-template <typename T, SparseMatrix SpMat>
-void to_explicit_buffer(const SpMat &a, T *mat_a, Layout layout) {
-    using sint_t = typename SpMat::index_t;
-    constexpr bool is_coo = std::is_same_v<SpMat, COOMatrix<T, sint_t>>;
-    constexpr bool is_csc = std::is_same_v<SpMat, CSCMatrix<T, sint_t>>;
-    constexpr bool is_csr = std::is_same_v<SpMat, CSRMatrix<T, sint_t>>;
-    if constexpr (is_coo) {
-        RandBLAS::sparse_data::coo::coo_to_dense(a, layout, mat_a);
-    } else if constexpr (is_csc) {
-        RandBLAS::sparse_data::csc::csc_to_dense(a, layout, mat_a);
-    } else if constexpr (is_csr) {
-        RandBLAS::sparse_data::csr::csr_to_dense(a, layout, mat_a);
-    } else {
-        randblas_require(false);
-    }
-    return;
-}
-
-template <typename T>
-void to_explicit_buffer(const SparseSkOp<T> &a, T *mat_a, Layout layout) {
-    if (a.nnz < 0) {
-        SparseSkOp<T> a_shallowcopy(a.dist, a.seed_state);
-        fill_sparse(a_shallowcopy);
-        auto a_coo = RandBLAS::sparse::coo_view_of_skop(a_shallowcopy);
-        to_explicit_buffer(a_coo, mat_a, layout);
-    } else {
-        auto a_coo = RandBLAS::sparse::coo_view_of_skop(a);
-        to_explicit_buffer(a_coo, mat_a, layout); 
-    }
-    return;
-}
-
-template <typename T>
-void to_explicit_buffer(const DenseSkOp<T> &s, T *mat_s, Layout layout) {
-    auto n_rows = s.dist.n_rows;
-    auto n_cols = s.dist.n_cols;
-    auto [stride_row, stride_col] = layout_to_strides(layout, n_rows, n_cols);
-    #define MAT_S(_i, _j) mat_s[(_i) * stride_row + (_j) * stride_col ]
-
-    // for some reason we prefer to make a copy rather than pass-by-value.
-    DenseSkOp<T> s_copy(s.dist, s.seed_state);
-    RandBLAS::fill_dense(s_copy);
-    auto [buff_stride_row, buff_stride_col] = layout_to_strides(s_copy.layout, n_rows, n_cols);
-    #define BUFF(_i, _j) s_copy.buff[(_i) * buff_stride_row + (_j) * buff_stride_col]
-
-    for (int64_t i = 0; i < n_rows; ++i) {
-        for (int64_t j = 0; j < n_cols; ++j) {
-            MAT_S(i, j) = BUFF(i, j);
-        }
-    }
-    return;
-}
-
-
-// MARK:      Multiply from the LEFT
-////////////////////////////////////////////////////////////////////////
-////////////////////////////////////////////////////////////////////////
-
-
-template <typename T>
-void left_apply(Layout layout, Op opS, Op opA, int64_t d, int64_t n, int64_t m, T alpha, SparseSkOp<T> &S, int64_t S_ro, int64_t S_co, const T *A, int64_t lda, T beta, T *B, int64_t ldb, int threads = 0) {
-    #if defined (RandBLAS_HAS_OpenMP)
-        int orig_threads = omp_get_num_threads();
-        if (threads > 0)
-            omp_set_num_threads(threads);
-    #else
-        UNUSED(threads);
-    #endif
-    RandBLAS::sparse::lskges(layout, opS, opA, d, n, m, alpha, S, S_ro, S_co, A, lda, beta, B, ldb);
-    #if defined (RandBLAS_HAS_OpenMP)
-        omp_set_num_threads(orig_threads);
-    #endif
-    return;
-}
-
-template <typename T>
-void left_apply(Layout layout, Op opS, Op opA, int64_t d, int64_t n, int64_t m, T alpha, DenseSkOp<T> &S, int64_t S_ro, int64_t S_co, const T *A, int64_t lda, T beta, T *B, int64_t ldb, int threads = 0) {
-    #if defined (RandBLAS_HAS_OpenMP)
-        int orig_threads = omp_get_num_threads();
-        if (threads > 0)
-            omp_set_num_threads(threads);
-    #else
-        UNUSED(threads);
-    #endif
-    RandBLAS::dense::lskge3(layout, opS, opA, d, n, m, alpha, S, S_ro, S_co, A, lda, beta, B, ldb);
-    #if defined (RandBLAS_HAS_OpenMP)
-        omp_set_num_threads(orig_threads);
-    #endif
-    return;
-}
-
-template <typename T, SparseMatrix SpMat>
-void left_apply(Layout layout, Op opS, Op opA, int64_t d, int64_t n, int64_t m, T alpha, const SpMat &S, int64_t S_ro, int64_t S_co, const T *A, int64_t lda, T beta, T *B, int64_t ldb, int threads = 0) {
-    #if defined (RandBLAS_HAS_OpenMP)
-        int orig_threads = omp_get_num_threads();
-        if (threads > 0)
-            omp_set_num_threads(threads);
-    #else
-        UNUSED(threads);
-    #endif
-    RandBLAS::sparse_data::left_spmm(layout, opS, opA, d, n, m, alpha, S, S_ro, S_co, A, lda, beta, B, ldb);
-    #if defined (RandBLAS_HAS_OpenMP)
-        omp_set_num_threads(orig_threads);
-    #endif
-    return;
-}
-
-template <typename T, typename LinOp>
-void reference_left_apply(
-    Layout layout, Op transS, Op transA, int64_t d, int64_t n, int64_t m, T alpha, LinOp &S, int64_t S_ro, int64_t S_co, const T *A, int64_t lda, T beta, T *B, T *E, int64_t ldb
-){
-    randblas_require(d > 0);
-    randblas_require(m > 0);
-    randblas_require(n > 0);
-
-    // Dimensions of mat(A), rather than op(mat(A))
-    auto [rows_S, cols_S] = dimensions(S);
-    auto [rows_mat_A   , cols_mat_A   ] = RandBLAS::dims_before_op(m, n, transA);
-    auto [rows_submat_S, cols_submat_S] = RandBLAS::dims_before_op(d, m, transS);
-
-    // Sanity checks on dimensions and strides
-    int64_t lds, s_row_stride, s_col_stride, pos, size_A, size_B;
-    if (layout == Layout::ColMajor) {
-        lds = rows_S;
-        pos = S_ro + lds * S_co;
-        randblas_require(lds >= rows_submat_S);
-        randblas_require(lda >= rows_mat_A);
-        randblas_require(ldb >= d);
-        size_A = lda * (cols_mat_A - 1) + rows_mat_A;;
-        size_B = ldb * (n - 1) + d;
-        s_row_stride = 1;
-        s_col_stride = lds;
-    } else {
-        lds = cols_S;
-        pos = S_ro * lds + S_co;
-        randblas_require(lds >= cols_submat_S);
-        randblas_require(lda >= cols_mat_A);
-        randblas_require(ldb >= n);
-        size_A = lda * (rows_mat_A - 1) + cols_mat_A;
-        size_B = ldb * (d - 1) + n;
-        s_row_stride = lds;
-        s_col_stride = 1;
-    }
-
-    auto size_S = rows_S * cols_S;
-    std::vector<T> S_dense(size_S);
-    std::vector<T> S_dense_abs(size_S);
-    to_explicit_buffer(S, S_dense.data(), layout);
-    for (int64_t i = 0; i < rows_S; ++i) {
-        for (int64_t j = 0; j < cols_S; ++j) {
-            auto ell = i * s_row_stride + j * s_col_stride;
-            S_dense_abs[ell] = abs(S_dense[ell]);
-        }
-    }
-
-    // Compute the reference value
-    T* S_ptr = S_dense.data();
-    blas::gemm(layout, transS, transA, d, n, m,
-        alpha, &S_ptr[pos], lds, A, lda, beta, B, ldb
-    );
-
-    // Compute the matrix needed for componentwise error bounds.
-    std::vector<T> A_abs_vec(size_A);
-    T* A_abs = A_abs_vec.data();
-    for (int64_t i = 0; i < size_A; ++i)
-        A_abs[i] = abs(A[i]);
-    if (beta != 0.0) {
-        for (int64_t i = 0; i < size_B; ++i)
-            E[i] = abs(B[i]);
-    }
-    T eps = std::numeric_limits<T>::epsilon();
-    T err_alpha = (abs(alpha) * m) * (2 * eps);
-    T err_beta = abs(beta) * eps;
-    T* S_abs_ptr = S_dense_abs.data();
-    blas::gemm(layout, transS, transA, d, n, m,
-        err_alpha, &S_abs_ptr[pos], lds, A_abs, lda, err_beta, E, ldb
-    );
-    return;
-}
 
 template <typename T, typename LinOp>
 void test_left_apply_to_random(
@@ -280,12 +65,12 @@ void test_left_apply_to_random(
     int64_t lda = (is_colmajor) ? m : n;
     int64_t ldb = (is_colmajor) ? d : n;
 
-    // compute S*A. 
+    // compute S*A.
     left_apply<T>(
         layout, Op::NoTrans, Op::NoTrans,
         d, n, m,
         alpha, S, 0, 0, A.data(), lda,
-        beta, B0.data(), ldb, threads 
+        beta, B0.data(), ldb, threads
     );
 
     // compute expected result (B1) and allowable error (E)
@@ -328,13 +113,13 @@ void test_left_apply_submatrix_to_eye(
         d1, m1, m1,
         alpha, S0, S_ro, S_co,
         A.data(), lda,
-        beta, B.data(), ldb, threads   
+        beta, B.data(), ldb, threads
     );
 
     // Check the result
     T *expect = new T[d0 * m0];
     to_explicit_buffer(S0, expect, layout);
-    int64_t ld_expect = (is_colmajor) ? d0 : m0; 
+    int64_t ld_expect = (is_colmajor) ? d0 : m0;
     auto [row_stride_s, col_stride_s] = RandBLAS::layout_to_strides(layout, ld_expect);
     auto [row_stride_b, col_stride_b] = RandBLAS::layout_to_strides(layout, ldb);
     int64_t offset = row_stride_s * S_ro + col_stride_s * S_co;
@@ -375,7 +160,7 @@ void test_left_apply_transpose_to_eye(
         Op::NoTrans,
         d, m, m,
         1.0, S, 0, 0, A.data(), m,
-        0.0, B.data(), ldb, threads   
+        0.0, B.data(), ldb, threads
     );
 
     std::vector<T> S_dense(m * d, 0.0);
@@ -401,7 +186,7 @@ void test_left_apply_to_submatrix(
 
     auto A = std::get<0>(random_matrix<T>(m0, n0, RNGState(13)));
     auto [a_offset, lda] = offset_and_ldim(layout, m0, n0, A_ro, A_co);
-    T *A_ptr = &A.data()[a_offset]; 
+    T *A_ptr = &A.data()[a_offset];
     left_apply<T>(
         layout,
         Op::NoTrans,
@@ -409,7 +194,7 @@ void test_left_apply_to_submatrix(
         d, n, m,
         1.0, S, 0, 0,
         A_ptr, lda,
-        0.0, B0.data(), ldb, threads   
+        0.0, B0.data(), ldb, threads
     );
 
     std::vector<T> B1(d * n, 0.0);
@@ -448,7 +233,7 @@ void test_left_apply_to_transposed(
         d, n, m,
         1.0, S, 0, 0,
         At.data(), lda,
-        0.0, B0.data(), ldb, threads   
+        0.0, B0.data(), ldb, threads
     );
 
     std::vector<T> B1(d * n, 0.0);
@@ -471,87 +256,6 @@ void test_left_apply_to_transposed(
 // MARK:      Multiply from the RIGHT
 ////////////////////////////////////////////////////////////////////////
 ////////////////////////////////////////////////////////////////////////
-
-template <typename T>
-void right_apply(Layout layout, Op transA, Op transS, int64_t m, int64_t d, int64_t n, T alpha, const T *A, int64_t lda, const SparseSkOp<T> &S, int64_t S_ro, int64_t S_co, T beta, T *B, int64_t ldb, int threads) {
-    #if defined (RandBLAS_HAS_OpenMP)
-        int orig_threads = omp_get_num_threads();
-        if (threads > 0)
-            omp_set_num_threads(threads);
-    #else
-        UNUSED(threads);
-    #endif
-    RandBLAS::sparse::rskges(layout, transA, transS, m, d, n, alpha, A, lda, S, S_ro, S_co, beta, B, ldb);
-    #if defined (RandBLAS_HAS_OpenMP)
-        omp_set_num_threads(orig_threads);
-    #endif
-}
-
-template <typename T>
-void right_apply(Layout layout, Op transA, Op transS, int64_t m, int64_t d, int64_t n, T alpha, const T *A, int64_t lda, const DenseSkOp<T> &S, int64_t S_ro, int64_t S_co, T beta, T *B, int64_t ldb, int threads) {
-    #if defined (RandBLAS_HAS_OpenMP)
-        int orig_threads = omp_get_num_threads();
-        if (threads > 0)
-            omp_set_num_threads(threads);
-    #else
-        UNUSED(threads);
-    #endif
-    RandBLAS::dense::rskge3(layout, transA, transS, m, d, n, alpha, A, lda, S, S_ro, S_co, beta, B, ldb);
-    #if defined (RandBLAS_HAS_OpenMP)
-        omp_set_num_threads(orig_threads);
-    #endif
-}
-
-template <typename T, SparseMatrix SpMat>
-void right_apply(Layout layout, Op transA, Op transS, int64_t m, int64_t d, int64_t n, T alpha, const T *A, int64_t lda, const SpMat &S, int64_t S_ro, int64_t S_co, T beta, T *B, int64_t ldb, int threads) {
-    #if defined (RandBLAS_HAS_OpenMP)
-        int orig_threads = omp_get_num_threads();
-        if (threads > 0)
-            omp_set_num_threads(threads);
-    #else
-        UNUSED(threads);
-    #endif
-    RandBLAS::sparse_data::right_spmm(layout, transA, transS, m, d, n, alpha, A, lda, S, S_ro, S_co, beta, B, ldb);
-    #if defined (RandBLAS_HAS_OpenMP)
-        omp_set_num_threads(orig_threads);
-    #endif
-}
-
-template <typename T, typename LinOp>
-void reference_right_apply(
-    Layout layout, Op transA, Op transS, int64_t m, int64_t d, int64_t n, T alpha, const T *A, int64_t lda, const LinOp &S0, int64_t S_ro, int64_t S_co, T beta, T *B, T *E, int64_t ldb
-) { 
-    // Check dimensions of submat(S).
-    auto [submat_S_rows, submat_S_cols] = RandBLAS::dims_before_op(n, d, transS);
-    auto [rows_S, cols_S] = dimensions(S0);
-    randblas_require(submat_S_rows <= rows_S);
-    randblas_require(submat_S_cols <= cols_S);
-    // Check dimensions of mat(A).
-    auto [mat_A_rows, mat_A_cols] = RandBLAS::dims_before_op(m, n, transA);
-    if (layout == Layout::ColMajor) {
-        randblas_require(lda >= mat_A_rows);
-    } else {
-        randblas_require(lda >= mat_A_cols);
-    }
-    //
-    // Compute B = op(A) op(submat(S)) by left_apply. We start with the identity
-    //
-    //      B^T = op(submat(S))^T op(A)^T
-    //
-    // Then we interchange the operator "op" for op(A) and the operator (*)^T.
-    //
-    //      B^T = op(submat(S))^T op(A^T)
-    //
-    // We tell left_apply to process (B^T) and (A^T) in the opposite memory layout
-    // compared to the layout for (A, B).
-    // 
-    auto trans_transS = (transS == Op::NoTrans) ? Op::Trans : Op::NoTrans;
-    auto trans_layout = (layout == Layout::ColMajor) ? Layout::RowMajor : Layout::ColMajor;
-    reference_left_apply(
-        trans_layout, trans_transS, transA,
-        d, m, n, alpha, S0, S_ro, S_co, A, lda, beta, B, E, ldb
-    );
-}
 
 template <typename T, typename LinOp>
 void test_right_apply_to_random(
@@ -636,13 +340,13 @@ void test_right_apply_transpose_to_eye(
     bool is_colmajor = Layout::ColMajor == layout;
     int64_t ldb = (is_colmajor) ? n : d;
     int64_t lds = (is_colmajor) ? d : n;
-    
+
     right_apply<T>(layout, Op::NoTrans, Op::Trans, n, d, n, 1.0, A.data(), n, S, 0, 0, 0.0, B.data(), ldb, threads);
 
     std::vector<T> S_dense(n * d, 0.0);
     to_explicit_buffer(S, S_dense.data(), layout);
     test::comparison::matrices_approx_equal(
-        layout, Op::Trans, n, d, 
+        layout, Op::Trans, n, d,
         B.data(), ldb, S_dense.data(), lds,
         __PRETTY_FUNCTION__, __FILE__, __LINE__
     );
@@ -656,7 +360,7 @@ void test_right_apply_to_submatrix(
     auto [n, d] = dimensions(S);
     randblas_require(m0 >= m);
     randblas_require(n0 >= n);
-    
+
     std::vector<T> B0(m * d, 0.0);
     int64_t ldb = (layout == Layout::ColMajor) ? m : d;
 
@@ -680,7 +384,7 @@ void test_right_apply_to_submatrix(
         B0.data(), B1.data(), E.data(), d * m,
         __PRETTY_FUNCTION__, __FILE__, __LINE__
     );
-    
+
 }
 
 template <typename T, typename LinOp>
